requirements.txt
backcall==0.2.0
beautifulsoup4==4.10.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
contourpy==1.1.0
cryptography==36.0.0
cycler==0.11.0
decorator==5.1.1
dnspython==2.2.1
filelock==3.4.2
fonttools==4.42.0
glob2==0.7
idna==3.3
importlib-resources==6.0.0
ipython==7.31.1
jedi==0.18.1
Jinja2==2.11.3
joblib==1.3.1
kiwisolver==1.4.4
libarchive-c==2.9
MarkupSafe==2.0.1
matplotlib==3.7.2
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.21.2
opencv-python-headless==4.8.0.74
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
pkginfo==1.8.2
prompt-toolkit==3.0.20
psutil==5.8.0
ptyprocess==0.7.0
pycosat==0.6.3
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==21.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel-yaml-conda==0.15.100
scikit-learn==1.3.0
scipy==1.10.1
setuptools==58.0.4
six==1.16.0
soupsieve==2.3.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==1.11.0
torchelastic==0.2.2
torchtext==0.12.0
torchvision==0.12.0
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
wheel==0.37.1
zipp==3.16.2
